requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
pillow==11.1.0
pandas==2.2.3
pyyaml==6.0.3
tqdm==4.67.1